+++ src/Contract/Routing/Route.php
@@ -16,40 +16,40 @@ interface Route
      *
      * @return array<class-string, string>|callable|string
      */
-    public function handler();
+    public function getHandler();
 
     /**
      * Returns the HTTP method.
      *
      * @return string
      */
-    public function method();
+    public function getMethod();
 
     /**
      * Returns an array of middlewares.
      *
-     * @return array<int, \Zapheus\Contract\Http\Server\Middleware>
+     * @return array<integer, \Zapheus\Contract\Http\Server\Middleware>
      */
-    public function middlewares();
+    public function getMiddlewares();
 
     /**
      * Returns the parameters if any.
      *
      * @return array<string, string>
      */
-    public function parameters();
+    public function getParams();
 
     /**
      * Returns a regular expression from URI.
      *
      * @return string
      */
-    public function regex();
+    public function getRegex();
 
     /**
      * Returns the URI.
      *
      * @return string
      */
-    public function uri();
+    public function getUri();
 }